* 'master' of https://github.com/opfo/app: (160 commits)
  Fix minor bug in tag-loading.
  Add header for related tag in tag-browser.
  Deleted some code that wasn't used
  Change NSLog to DLog.
  Remove NSLog call.
  Fix crash when OPFQuestionsViewController was deallocated.
  Fix multiple references to the same file in Xcode.
  Fix laying out question metadata header view subviews.
  An user which is not logged in cannotpost questions, answers or comments anymore
  Fix app-crash when logging out after tab-move.
  Change font to avenir next in activity view.
  An user which is not logged in cannotpost questions, answers or comments anymore
  Change profile search view font to avenir next.
  Change footer, header lookz. and fix more issues.
  Add highlight background for user preview button.
  Add background image for user preview button.
  Change sidebar's gradient saturation.
  Score is updated in DB when user vote
  Fix a failing spec.
  Add gradient to sidebar of comment cell.
  ...

Conflicts:
	Code Stream/OPFCommentViewCell.xib

+++ Code Stream/UIColor+OPFHEX.m
@@ -10,7 +10,7 @@
 
 @implementation UIColor (OPFHEX)
 
-#define OPFHEXUIColorFromRGB(rgbValue) [UIColor colorWithRed:((float)((rgbValue & 0xFF0000) >> 16))/255.0 green:((float)((rgbValue & 0xFF00) >> 8))/255.0 blue:((float)(rgbValue & 0xFF))/255.0 alpha:1.0]
+#define OPFHEXUIColorFromRGB(rgbValue) [UIColor colorWithRed:((CGFloat)((rgbValue & 0xFF0000) >> 16))/255.f green:((CGFloat)((rgbValue & 0xFF00) >> 8))/255.f blue:((CGFloat)(rgbValue & 0xFF))/255.f alpha:1.f]
 
 + (instancetype)opf_colorWithHexString:(NSString *)hexString
 {
